+++ Source/swiftlint/Commands/Rules.swift
@@ -10,28 +10,13 @@ import Foundation
 import SwiftLintFramework
 import SwiftyTextTable
 
-enum RuleEnablementOptions: String, EnumerableFlag {
-    case enabled, disabled
-
-    static func name(for value: RuleEnablementOptions) -> NameSpecification {
-        return .shortAndLong
-    }
-
-    static func help(for value: RuleEnablementOptions) -> ArgumentHelp? {
-        return "Only show \(value.rawValue) rules"
-    }
-}
-
 extension SwiftLint {
     struct Rules: ParsableCommand {
         static let configuration = CommandConfiguration(abstract: "Display the list of rules and their identifiers")
 
         @Option(help: "The path to a SwiftLint configuration file")
         var config: String?
-        @Flag(exclusivity: .exclusive)
-        var ruleEnablement: RuleEnablementOptions?
-        @Flag(name: .shortAndLong, help: "Only display correctable rules")
-        var correctable = false
+        @OptionGroup var rulesFilterOptions: RulesFilterOptions
         @Flag(name: .shortAndLong, help: "Display full configuration details")
         var verbose = false
         @Argument(help: "The rule identifier to display description for")
@@ -48,34 +33,11 @@ extension SwiftLint {
             }
 
             let configuration = Configuration(configurationFiles: [config].compactMap({ $0 }))
-            let rules = ruleList(configuration: configuration)
+            let rulesFilter = RulesFilter(enabledRules: configuration.rules)
+            let rules = rulesFilter.getRules(excluding: .excludingOptions(byCommandLineOptions: rulesFilterOptions))
             let table = TextTable(ruleList: rules, configuration: configuration, verbose: verbose)
             print(table.render())
             ExitHelper.successfullyExit()
-        }
-
-        private func ruleList(configuration: Configuration) -> RuleList {
-            guard ruleEnablement != nil || correctable else {
-                return primaryRuleList
-            }
-
-            let filtered: [Rule.Type] = primaryRuleList.list.compactMap { ruleID, ruleType in
-                let configuredRule = configuration.rules.first { rule in
-                    return type(of: rule).description.identifier == ruleID
-                }
-
-                if ruleEnablement == .enabled && configuredRule == nil {
-                    return nil
-                } else if ruleEnablement == .disabled && configuredRule != nil {
-                    return nil
-                } else if correctable && !(configuredRule is CorrectableRule) {
-                    return nil
-                }
-
-                return ruleType
-            }
-
-            return RuleList(rules: filtered)
         }
     }
 }