+++ CHANGELOG.md
@@ -32,6 +32,9 @@
   [JP Simard](https://github.com/jpims)
   [#4031](https://github.com/realm/SwiftLint/issues/4031)
 
+* Add ability to filter rules for `generate-docs` subcommand.  
+  [kattouf](https://github.com/kattouf)
+
 * Add new `excludes_trivial_init` configuration for `missing_docs` rule
   to exclude initializers without any parameters.  
   [Marcelo Fabri](https://github.com/marcelofabri)